Test Final

+++ Tour_du_lich/Controllers/DiaDiemController.cs
@@ -59,8 +59,17 @@ namespace Tour_du_lich.Controllers
         {
             try
             {
-                ddDao.Delete(id);
-                string code = Constants.SUCCESS;
+                string code;
+                if(ddDao.ExistIdInAnotherTable(id))
+                {
+                    code = Constants.EXISTS_FOREIGN_KEY;
+                }
+                else
+                {
+                    ddDao.Delete(id);
+                    code = Constants.SUCCESS;
+                }
+
                 return Json(new { Code = code, JsonRequestBehavior.AllowGet });
             }
             catch (Exception ex)
